requirements.txt
Flask>=2.0
google-generativeai>=0.3
rank_bm25>=0.2
sentence-transformers>=2.2
numpy>=1.20
PyMuPDF>=1.18 # fitz
python-docx>=0.8
python-dotenv>=0.19 # Optional, for loading .env files
# Add other dependencies if you used them, e.g., docx2txt
